+++ ChangeLog.md
@@ -2,6 +2,7 @@
 <summary> 3.10.3 - 2024-09-15 </summary>
 
 - Implemented "Midas Touch" tier upgrade which increases the value of scrap found in the moons.
+- Implemented "Quick Hands" tier upgrade which increases the player's interaction speed.
 - Changed upgrade display to show the price number red if the players do not have enough credits to purchase it.
 - Changed upgrade store to show first pick between shared and individual upgrades if any are configured as so and show the list of upgrades when selected.
   - If all enabled upgrades are either individual or shared, it will do the same behaviour as before.